Hold menu easing tick rate as float
Convert the tick-rate denominator to float before reading the elapsed tick value so the easing helper follows the target's floating-point lifetime and division schedule.

Constraint: Preserve the corrected non-strict easing boundaries.
Confidence: high
Scope-risk: narrow
Directive: Keep the denominator as an early f32 conversion unless exact scoring proves otherwise.
Tested: exact menuCB_r50_800693A4_suffix build; focused objdiff 53.522728% to 68.659090%; git diff --check.
Not-tested: runtime menu transition timing.

+++ src/game/menu/menuCB_range_80062948.c
@@ -1613,12 +1613,12 @@ static inline void menuCBPokemonEntryAdvancePositions(void)
     f32 elapsed;
     s32 player;
     s32 component;
-    s32 denominator;
+    f32 denominator;
     s32 numerator;
 
-    denominator = fn_800D37CC();
+    denominator = (f32)fn_800D37CC();
     numerator = fn_800D3088();
-    elapsed = (f32)numerator / (f32)denominator;
+    elapsed = (f32)numerator / denominator;
     *(f32*)&lbl_803A9F08[0xCD88] = elapsed;
     for (player = 0; player < 4; player++) {
         current = (f32*)&lbl_803A9F08[0xCD8C + player * 0x30];
